examples/embedding/requirements.txt
pytorch
nltk
scikit-learn
pandas
h5py
cython
tqdm
numpy
